requirements.txt
# Learnify - AI Teaching Assistant Dependencies
# Core
anthropic>=0.39.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0

# PDF Generation
reportlab>=4.0.0

# Production server
gunicorn>=21.0.0

# Utilities
python-dateutil>=2.8.0
